+++ src/niggli.h
@@ -35,6 +35,8 @@
 #ifndef __NIGGLI_H__
 #define __NIGGLI_H__
 
+#include "base.h"
+
 #define NIGGLI_MAJOR_VERSION 0
 #define NIGGLI_MINOR_VERSION 1
 #define NIGGLI_MICRO_VERSION 2
@@ -42,7 +44,7 @@
 int niggli_get_major_version(void);
 int niggli_get_minor_version(void);
 int niggli_get_micro_version(void);
-int niggli_reduce(double* lattice_, const double eps_,
-                  const int aperiodic_axis);
+SPG_API_TEST int niggli_reduce(double* lattice_, const double eps_,
+                               const int aperiodic_axis);
 
 #endif